editor/modeling: CubeGrid free-height boxes — continuous stretch, not cube stacks
The extrude height was quantized to whole cells, so raising height added another
cube. Reworked the model from voxel cells to grid-aligned BOXES with a free,
continuous world-space height:

- LMB drops a W×D grid footprint (snaps to the grid); drag along the face normal
  STRETCHES the box to any height (H readout in world units) — a smooth
  deformation, never a stack of unit cubes
- ModelingState.Height is now a float (world units); a plain click makes a box of
  that height; editing it (panel DragFloat, E/Q, arrows) re-deforms the last box
  live before Accept
- Targeting raycasts box AABBs -> stack on any face, else the ground; Shift+LMB /
  RMB removes a whole box
- RegenMesh bakes each box as a stretched cube; solid grey checkerboard overlay +
  blue live preview

+++ Editor/Source/Editor/Core/Selection/ModelingState.hpp
@@ -33,8 +33,8 @@ namespace Desert::Editor::Core
         float  CellSize   = 1.0f;   // grid step
         int    BrushW     = 1;      // CubeGrid paint footprint (cells) across the two in-plane axes
         int    BrushD     = 1;
-        int    Height     = 1;      // CubeGrid extrude height (cells) — drag sets it; editing here / E-Q
-                                    // re-extrudes the last column live before Accept
+        float  Height     = 1.0f;   // CubeGrid extrude height (WORLD units, free/continuous) — drag sets it;
+                                    // editing here / E-Q re-deforms the last box live before Accept
         bool   ReqAccept  = false;  // one-shot: commit the blockout, start a fresh one
         bool   ReqCancel  = false;  // one-shot: delete the in-progress blockout
         bool   ReqClear   = false;  // one-shot: clear the cells (keep editing)
